Move wait_for_pid_to_exit to test_utils.py.
